pyproject.toml
[project]
name = "txtextracteval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Alex Kovac", email = "5686693+kovaca@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.29.0",
    "openai>=1.35.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "python-levenshtein>=0.27.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "sentencepiece>=0.2.0",
    "torch>=2.7.0",
    "transformers>=4.51.3",
    "tabulate>=0.9.0",
    "google-genai>=1.13.0",
    "google-api-core>=2.24.2",
    "litellm>=1.67.5",
    "easyocr>=1.7.1",
    "torchvision>=0.18.0",
    "seaborn>=0.13.2",
]

[project.scripts]
# txtextracteval = "txtextracteval:main"
txtextracteval = "txtextracteval.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0",
]
